Move previous configuration file into its own space.
